+++ SignNow.Net/Model/Responses/CreateDocumentFromTemplateResponse.cs
@@ -1,0 +1,16 @@
+using Newtonsoft.Json;
+
+namespace SignNow.Net.Model.Responses
+{
+    /// <summary>
+    /// Represents response from SignNow API for Create Document from Template request.
+    /// </summary>
+    public class CreateDocumentFromTemplateResponse
+    {
+        /// <summary>
+        /// Identity of new Document.
+        /// </summary>
+        [JsonProperty("id")]
+        public string Id { get; set; }
+    }
+}